+++ src/clojure_lsp/kondo.clj
@@ -1,6 +1,7 @@
 (ns clojure-lsp.kondo
   (:require
    [clj-kondo.core :as kondo]
+   [clj-kondo.impl.core :as kondo.impl]
    [clojure-lsp.feature.diagnostics :as f.diagnostic]
    [clojure-lsp.shared :as shared]
    [clojure.core.async :as async]
@@ -156,3 +157,10 @@
         (when-not (string/blank? (str err))
           (log/error (str err)))
         result))))
+
+(defn config-hash
+  [project-root]
+  (-> project-root
+      (io/file ".clj-kondo")
+      kondo/resolve-config
+      kondo/config-hash))